Remove local amclient imports
This commit removes local references to amclient and elects to import
the PyPi package. Tests have been updated to work and VCR Cassettes
have been removed.
