--- fixtures/vcr_cassettes/aip2dips_dip.yaml
@@ -1,55 +1,0 @@
-interactions:
-- request:
-    body: null
-    headers:
-      Accept: ['*/*']
-      Accept-Encoding: ['gzip, deflate']
-      Connection: [keep-alive]
-      User-Agent: [python-requests/2.11.1]
-    method: GET
-    uri: http://192.168.168.192:8000/api/v2/file/?package_type=DIP&username=test&api_key=5de62f6f4817f903dcfac47fa5cffd44685a2cf2
-  response:
-    body: {string: '{"meta": {"limit": 1, "next": "/api/v2/file/?username=test&api_key=5de62f6f4817f903dcfac47fa5cffd44685a2cf2&limit=1&offset=1&package_type=DIP",
-        "offset": 0, "previous": null, "total_count": 2}, "objects": [{"current_full_path":
-        "/var/archivematica/sharedDirectory/www/DIPsStore/7e49/afa4/116b/4650/8bbb/9341/906b/db21/make_dips-99bb20ee-69c6-43d0-acf0-c566020357d2",
-        "current_location": "/api/v2/location/5150d0e2-f643-404c-aab5-cf7ebcea05bf/",
-        "current_path": "7e49/afa4/116b/4650/8bbb/9341/906b/db21/make_dips-99bb20ee-69c6-43d0-acf0-c566020357d2",
-        "origin_pipeline": "/api/v2/pipeline/a702cea2-9666-42e4-9784-6881e3fb5f67/",
-        "package_type": "DIP", "resource_uri": "/api/v2/file/7e49afa4-116b-4650-8bbb-9341906bdb21/",
-        "size": 211240, "status": "UPLOADED", "uuid": "7e49afa4-116b-4650-8bbb-9341906bdb21"}]}'}
-    headers:
-      Cache-Control: [no-cache]
-      Connection: [keep-alive]
-      Content-Type: [application/json]
-      Date: ['Mon, 07 Nov 2016 21:17:45 GMT']
-      Server: [nginx/1.4.6 (Ubuntu)]
-      Vary: ['Accept, Cookie']
-      X-Frame-Options: [SAMEORIGIN]
-    status: {code: 200, message: OK}
-- request:
-    body: null
-    headers:
-      Accept: ['*/*']
-      Accept-Encoding: ['gzip, deflate']
-      Connection: [keep-alive]
-      User-Agent: [python-requests/2.11.1]
-    method: GET
-    uri: http://192.168.168.192:8000/api/v2/file/?username=test&api_key=5de62f6f4817f903dcfac47fa5cffd44685a2cf2&limit=1&offset=1&package_type=DIP
-  response:
-    body: {string: '{"meta": {"limit": 1, "next": null, "offset": 1, "previous": "/api/v2/file/?username=test&api_key=5de62f6f4817f903dcfac47fa5cffd44685a2cf2&package_type=DIP&limit=1&offset=0",
-        "total_count": 2}, "objects": [{"current_full_path": "/var/archivematica/sharedDirectory/www/DIPsStore/c0e3/7bab/e51e/482d/a066/a277/330d/e9a7/make_dips_2-721b98b9-b894-4cfb-80ab-624e52263300",
-        "current_location": "/api/v2/location/5150d0e2-f643-404c-aab5-cf7ebcea05bf/",
-        "current_path": "c0e3/7bab/e51e/482d/a066/a277/330d/e9a7/make_dips_2-721b98b9-b894-4cfb-80ab-624e52263300",
-        "origin_pipeline": "/api/v2/pipeline/a702cea2-9666-42e4-9784-6881e3fb5f67/",
-        "package_type": "DIP", "resource_uri": "/api/v2/file/c0e37bab-e51e-482d-a066-a277330de9a7/",
-        "size": 211400, "status": "UPLOADED", "uuid": "c0e37bab-e51e-482d-a066-a277330de9a7"}]}'}
-    headers:
-      Cache-Control: [no-cache]
-      Connection: [keep-alive]
-      Content-Type: [application/json]
-      Date: ['Mon, 07 Nov 2016 21:17:45 GMT']
-      Server: [nginx/1.4.6 (Ubuntu)]
-      Vary: ['Accept, Cookie']
-      X-Frame-Options: [SAMEORIGIN]
-    status: {code: 200, message: OK}
-version: 1
