--- fixtures/vcr_cassettes/aips_aips.yaml
@@ -1,55 +1,0 @@
-interactions:
-- request:
-    body: null
-    headers:
-      Accept: ['*/*']
-      Accept-Encoding: ['gzip, deflate']
-      Connection: [keep-alive]
-      User-Agent: [python-requests/2.11.1]
-    method: GET
-    uri: http://192.168.168.192:8000/api/v2/file/?username=test&api_key=5de62f6f4817f903dcfac47fa5cffd44685a2cf2&package_type=AIP
-  response:
-    body: {string: '{"meta": {"limit": 1, "next": "/api/v2/file/?username=test&api_key=5de62f6f4817f903dcfac47fa5cffd44685a2cf2&limit=1&offset=1&package_type=AIP",
-        "offset": 0, "previous": null, "total_count": 2}, "objects": [{"current_full_path":
-        "/var/archivematica/sharedDirectory/www/AIPsStore/979c/ce65/2a6f/407f/a49c/1bcf/13bd/8571/make-979cce65-2a6f-407f-a49c-1bcf13bd8571.7z",
-        "current_location": "/api/v2/location/91b917bb-57ea-4cca-8c16-f6b598713f93/",
-        "current_path": "979c/ce65/2a6f/407f/a49c/1bcf/13bd/8571/make-979cce65-2a6f-407f-a49c-1bcf13bd8571.7z",
-        "origin_pipeline": "/api/v2/pipeline/a702cea2-9666-42e4-9784-6881e3fb5f67/",
-        "package_type": "AIP", "resource_uri": "/api/v2/file/979cce65-2a6f-407f-a49c-1bcf13bd8571/",
-        "size": 24714, "status": "UPLOADED", "uuid": "979cce65-2a6f-407f-a49c-1bcf13bd8571"}]}'}
-    headers:
-      Cache-Control: [no-cache]
-      Connection: [keep-alive]
-      Content-Type: [application/json]
-      Date: ['Mon, 07 Nov 2016 20:44:12 GMT']
-      Server: [nginx/1.4.6 (Ubuntu)]
-      Vary: ['Accept, Cookie']
-      X-Frame-Options: [SAMEORIGIN]
-    status: {code: 200, message: OK}
-- request:
-    body: null
-    headers:
-      Accept: ['*/*']
-      Accept-Encoding: ['gzip, deflate']
-      Connection: [keep-alive]
-      User-Agent: [python-requests/2.11.1]
-    method: GET
-    uri: http://192.168.168.192:8000/api/v2/file/?username=test&api_key=5de62f6f4817f903dcfac47fa5cffd44685a2cf2&limit=1&offset=1&package_type=AIP
-  response:
-    body: {string: '{"meta": {"limit": 1, "next": null, "offset": 1, "previous": "/api/v2/file/?username=test&api_key=5de62f6f4817f903dcfac47fa5cffd44685a2cf2&package_type=AIP&limit=1&offset=0",
-        "total_count": 2}, "objects": [{"current_full_path": "/var/archivematica/sharedDirectory/www/AIPsStore/3500/aee0/08ca/40ff/8d2d/9fe9/a2c3/ae3b/itbetter-3500aee0-08ca-40ff-8d2d-9fe9a2c3ae3b.7z",
-        "current_location": "/api/v2/location/91b917bb-57ea-4cca-8c16-f6b598713f93/",
-        "current_path": "3500/aee0/08ca/40ff/8d2d/9fe9/a2c3/ae3b/itbetter-3500aee0-08ca-40ff-8d2d-9fe9a2c3ae3b.7z",
-        "origin_pipeline": "/api/v2/pipeline/a702cea2-9666-42e4-9784-6881e3fb5f67/",
-        "package_type": "AIP", "resource_uri": "/api/v2/file/3500aee0-08ca-40ff-8d2d-9fe9a2c3ae3b/",
-        "size": 8991980, "status": "UPLOADED", "uuid": "3500aee0-08ca-40ff-8d2d-9fe9a2c3ae3b"}]}'}
-    headers:
-      Cache-Control: [no-cache]
-      Connection: [keep-alive]
-      Content-Type: [application/json]
-      Date: ['Mon, 07 Nov 2016 20:44:12 GMT']
-      Server: [nginx/1.4.6 (Ubuntu)]
-      Vary: ['Accept, Cookie']
-      X-Frame-Options: [SAMEORIGIN]
-    status: {code: 200, message: OK}
-version: 1
