--- fixtures/vcr_cassettes/approve_non_existing_transfer.yaml
@@ -1,25 +1,0 @@
-interactions:
-- request:
-    body: !!python/unicode directory=approve_1&type=standard
-    headers:
-      Accept: ['*/*']
-      Accept-Encoding: ['gzip, deflate']
-      Authorization: [!!python/unicode 'ApiKey test:test']
-      Connection: [keep-alive]
-      Content-Length: ['33']
-      Content-Type: [application/x-www-form-urlencoded]
-      User-Agent: [python-requests/2.18.4]
-    method: POST
-    uri: http://192.168.168.192/api/transfer/approve/
-  response:
-    body: {string: !!python/unicode '{"message": "Unable to find unapproved transfer
-        directory.", "error": true}'}
-    headers:
-      connection: [keep-alive]
-      content-language: [en]
-      content-type: [application/json]
-      date: ['Sun, 08 Apr 2018 07:02:56 GMT']
-      server: [nginx/1.12.2]
-      vary: ['Accept-Language, Cookie']
-    status: {code: 500, message: INTERNAL SERVER ERROR}
-version: 1
