--- fixtures/vcr_cassettes/close_completed_ingests_ingests.yaml
@@ -1,62 +1,0 @@
-interactions:
-- request:
-    body: null
-    headers:
-      Accept: ['*/*']
-      Accept-Encoding: ['gzip, deflate']
-      Connection: [keep-alive]
-      User-Agent: [python-requests/2.18.1]
-    method: GET
-    uri: http://192.168.168.192/api/ingest/completed?username=test&api_key=3c23b0361887ace72b9d42963d9acbdf06644673
-  response:
-    body: {string: '{"message": "Fetched completed ingests successfully.", "results":
-        ["2d75323c-70a0-4d5f-a8d0-762e729fc2b9", "57d7faff-c397-4485-9035-6eaeb5c35636"]}'}
-    headers:
-      Connection: [keep-alive]
-      Content-Language: [en]
-      Content-Type: [application/json]
-      Date: ['Mon, 07 May 2018 07:52:43 GMT']
-      Server: [nginx/1.12.2]
-      Vary: ['Accept-Language, Cookie']
-    status: {code: 200, message: OK}
-- request:
-    body: null
-    headers:
-      Accept: ['*/*']
-      Accept-Encoding: ['gzip, deflate']
-      Connection: [keep-alive]
-      Content-Length: ['0']
-      User-Agent: [python-requests/2.18.1]
-    method: DELETE
-    uri: http://192.168.168.192/api/ingest/2d75323c-70a0-4d5f-a8d0-762e729fc2b9/delete/?username=test&api_key=3c23b0361887ace72b9d42963d9acbdf06644673
-  response:
-    body: {string: '{"removed": true}'}
-    headers:
-      Connection: [keep-alive]
-      Content-Language: [en]
-      Content-Type: [application/json]
-      Date: ['Mon, 07 May 2018 07:52:43 GMT']
-      Server: [nginx/1.12.2]
-      Vary: ['Accept-Language, Cookie']
-    status: {code: 200, message: OK}
-- request:
-    body: null
-    headers:
-      Accept: ['*/*']
-      Accept-Encoding: ['gzip, deflate']
-      Connection: [keep-alive]
-      Content-Length: ['0']
-      User-Agent: [python-requests/2.18.1]
-    method: DELETE
-    uri: http://192.168.168.192/api/ingest/57d7faff-c397-4485-9035-6eaeb5c35636/delete/?username=test&api_key=3c23b0361887ace72b9d42963d9acbdf06644673
-  response:
-    body: {string: '{"removed": true}'}
-    headers:
-      Connection: [keep-alive]
-      Content-Language: [en]
-      Content-Type: [application/json]
-      Date: ['Mon, 07 May 2018 07:52:44 GMT']
-      Server: [nginx/1.12.2]
-      Vary: ['Accept-Language, Cookie']
-    status: {code: 200, message: OK}
-version: 1
