--- fixtures/vcr_cassettes/close_completed_transfers_no_transfers.yaml
@@ -1,22 +1,0 @@
-interactions:
-- request:
-    body: null
-    headers:
-      Accept: ['*/*']
-      Accept-Encoding: ['gzip, deflate']
-      Connection: [keep-alive]
-      User-Agent: [python-requests/2.11.1]
-    method: GET
-    uri: http://192.168.168.192/api/transfer/completed?api_key=3c23b0361887ace72b9d42963d9acbdf06644673&username=test
-  response:
-    body: {string: '{"message": "Fetched completed transfers successfully.", "results":
-        []}'}
-    headers:
-      Connection: [Keep-Alive]
-      Content-Type: [application/json]
-      Date: ['Mon, 07 Nov 2016 19:08:04 GMT']
-      Keep-Alive: ['timeout=5, max=100']
-      Server: [Apache/2.4.7 (Ubuntu)]
-      Vary: [Cookie]
-    status: {code: 200, message: OK}
-version: 1
