--- fixtures/vcr_cassettes/close_completed_transfers_transfers.yaml
@@ -1,62 +1,0 @@
-interactions:
-- request:
-    body: null
-    headers:
-      Accept: ['*/*']
-      Accept-Encoding: ['gzip, deflate']
-      Connection: [keep-alive]
-      User-Agent: [python-requests/2.18.1]
-    method: GET
-    uri: http://192.168.168.192/api/transfer/completed?username=test&api_key=3c23b0361887ace72b9d42963d9acbdf06644673
-  response:
-    body: {string: '{"message": "Fetched completed transfers successfully.", "results":
-        ["9bc0b1c7-658f-46d4-9a6f-4a282e8a8ee5", "d7bd50b5-6473-4e9f-8555-8515e55d0a16"]}'}
-    headers:
-      Connection: [keep-alive]
-      Content-Language: [en]
-      Content-Type: [application/json]
-      Date: ['Mon, 07 May 2018 08:00:02 GMT']
-      Server: [nginx/1.12.2]
-      Vary: ['Accept-Language, Cookie']
-    status: {code: 200, message: OK}
-- request:
-    body: null
-    headers:
-      Accept: ['*/*']
-      Accept-Encoding: ['gzip, deflate']
-      Connection: [keep-alive]
-      Content-Length: ['0']
-      User-Agent: [python-requests/2.18.1]
-    method: DELETE
-    uri: http://192.168.168.192/api/transfer/9bc0b1c7-658f-46d4-9a6f-4a282e8a8ee5/delete/?username=test&api_key=3c23b0361887ace72b9d42963d9acbdf06644673
-  response:
-    body: {string: '{"removed": true}'}
-    headers:
-      Connection: [keep-alive]
-      Content-Language: [en]
-      Content-Type: [application/json]
-      Date: ['Mon, 07 May 2018 08:00:02 GMT']
-      Server: [nginx/1.12.2]
-      Vary: ['Accept-Language, Cookie']
-    status: {code: 200, message: OK}
-- request:
-    body: null
-    headers:
-      Accept: ['*/*']
-      Accept-Encoding: ['gzip, deflate']
-      Connection: [keep-alive]
-      Content-Length: ['0']
-      User-Agent: [python-requests/2.18.1]
-    method: DELETE
-    uri: http://192.168.168.192/api/transfer/d7bd50b5-6473-4e9f-8555-8515e55d0a16/delete/?username=test&api_key=3c23b0361887ace72b9d42963d9acbdf06644673
-  response:
-    body: {string: '{"removed": true}'}
-    headers:
-      Connection: [keep-alive]
-      Content-Language: [en]
-      Content-Type: [application/json]
-      Date: ['Mon, 07 May 2018 08:00:02 GMT']
-      Server: [nginx/1.12.2]
-      Vary: ['Accept-Language, Cookie']
-    status: {code: 200, message: OK}
-version: 1
