--- fixtures/vcr_cassettes/completed_ingests_no_ingests.yaml
@@ -1,22 +1,0 @@
-interactions:
-- request:
-    body: null
-    headers:
-      Accept: ['*/*']
-      Accept-Encoding: ['gzip, deflate']
-      Connection: [keep-alive]
-      User-Agent: [python-requests/2.7.0 CPython/2.7.6 Linux/3.13.0-87-generic]
-    method: GET
-    uri: http://192.168.168.192/api/ingest/completed?username=test&api_key=3c23b0361887ace72b9d42963d9acbdf06644673
-  response:
-    body: {string: !!python/unicode '{"message": "Fetched completed ingests successfully.",
-        "results": []}'}
-    headers:
-      connection: [Keep-Alive]
-      content-type: [application/json]
-      date: ['Mon, 02 Jan 2017 21:25:28 GMT']
-      keep-alive: ['timeout=5, max=100']
-      server: [Apache/2.4.7 (Ubuntu)]
-      vary: [Cookie]
-    status: {code: 200, message: OK}
-version: 1
