--- fixtures/vcr_cassettes/completed_transfers_bad_key.yaml
@@ -1,21 +1,0 @@
-interactions:
-- request:
-    body: null
-    headers:
-      Accept: ['*/*']
-      Accept-Encoding: ['gzip, deflate']
-      Connection: [keep-alive]
-      User-Agent: [python-requests/2.11.1]
-    method: GET
-    uri: http://192.168.168.192/api/transfer/completed?api_key=bad+api+key&username=test
-  response:
-    body: {string: '{"message": "API key not valid.", "error": true}'}
-    headers:
-      Connection: [Keep-Alive]
-      Content-Type: [application/json]
-      Date: ['Mon, 07 Nov 2016 19:10:03 GMT']
-      Keep-Alive: ['timeout=5, max=100']
-      Server: [Apache/2.4.7 (Ubuntu)]
-      Vary: [Cookie]
-    status: {code: 403, message: FORBIDDEN}
-version: 1
