--- fixtures/vcr_cassettes/completed_transfers_transfers.yaml
@@ -1,22 +1,0 @@
-interactions:
-- request:
-    body: null
-    headers:
-      Accept: ['*/*']
-      Accept-Encoding: ['gzip, deflate']
-      Connection: [keep-alive]
-      User-Agent: [python-requests/2.11.1]
-    method: GET
-    uri: http://192.168.168.192/api/transfer/completed?username=test&api_key=3c23b0361887ace72b9d42963d9acbdf06644673
-  response:
-    body: {string: '{"message": "Fetched completed transfers successfully.", "results":
-        ["26841c49-bed4-4db3-bf77-1bb0ac5db32c", "2f7bb26b-d1c7-484f-b41a-7e3f2fefa084"]}'}
-    headers:
-      Connection: [Keep-Alive]
-      Content-Type: [application/json]
-      Date: ['Mon, 07 Nov 2016 18:56:55 GMT']
-      Keep-Alive: ['timeout=5, max=100']
-      Server: [Apache/2.4.7 (Ubuntu)]
-      Vary: [Cookie]
-    status: {code: 200, message: OK}
-version: 1
