--- fixtures/vcr_cassettes/dips_no_dips.yaml
@@ -1,23 +1,0 @@
-interactions:
-- request:
-    body: null
-    headers:
-      Accept: ['*/*']
-      Accept-Encoding: ['gzip, deflate']
-      Connection: [keep-alive]
-      User-Agent: [python-requests/2.11.1]
-    method: GET
-    uri: http://192.168.168.192:8000/api/v2/file/?username=test&api_key=5de62f6f4817f903dcfac47fa5cffd44685a2cf2&package_type=DIP
-  response:
-    body: {string: '{"meta": {"limit": 1, "next": null, "offset": 0, "previous": null,
-        "total_count": 0}, "objects": []}'}
-    headers:
-      Cache-Control: [no-cache]
-      Connection: [keep-alive]
-      Content-Type: [application/json]
-      Date: ['Mon, 07 Nov 2016 20:51:18 GMT']
-      Server: [nginx/1.4.6 (Ubuntu)]
-      Vary: ['Accept, Cookie']
-      X-Frame-Options: [SAMEORIGIN]
-    status: {code: 200, message: OK}
-version: 1
