--- fixtures/vcr_cassettes/download_aip_fail.yaml
@@ -1,25 +1,0 @@
-interactions:
-- request:
-    body: null
-    headers:
-      Accept: ['*/*']
-      Accept-Encoding: ['gzip, deflate']
-      Connection: [keep-alive]
-      User-Agent: [python-requests/2.18.4]
-    method: GET
-    uri: http://192.168.168.192:8000/api/v2/file/bad-aip-uuid/download/?api_key=7021334bee4c9155c07e531608dd28a9d8039420&username=test
-  response:
-    body:
-      string: !!binary |
-        H4sIAAAAAAAAAwAAAP//Ckotzi8tSk5VKM8syVAIDfV0UUhKTNFNzCzQLS3NTFFIyU8tVsjLL1FI
-        rcgsLgEAAAD//wMA5u6dYS4AAAA=
-    headers:
-      connection: [keep-alive]
-      content-encoding: [gzip]
-      content-type: [text/html; charset=utf-8]
-      date: ['Thu, 16 Nov 2017 13:57:11 GMT']
-      server: [nginx/1.4.6 (Ubuntu)]
-      vary: ['Accept, Cookie']
-      x-frame-options: [SAMEORIGIN]
-    status: {code: 404, message: NOT FOUND}
-version: 1
