--- fixtures/vcr_cassettes/download_dip_dip.yaml
@@ -1,44 +1,0 @@
-interactions:
-- request:
-    body: null
-    headers:
-      Accept: ['*/*']
-      Accept-Encoding: ['gzip, deflate']
-      Connection: [keep-alive]
-      User-Agent: [python-requests/2.11.1]
-    method: GET
-    uri: http://192.168.168.192:8000/api/v2/file/c0e37bab-e51e-482d-a066-a277330de9a7/download/?api_key=5de62f6f4817f903dcfac47fa5cffd44685a2cf2&username=test
-  response:
-    body:
-      string: !!binary |
-        bWFrZV9kaXBzXzItNzIxYjk4YjktYjg5NC00Y2ZiLTgwYWItNjI0ZTUyMjYzMzAwLwAAAAAAAAAA
-        AAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAADAwMDA3NzYAMDAwMDUx
-        NQAwMDAwNTE1ADAwMDAwMDAwMDAwADEzMDEwMTY1MjUxADAyMjAyNgAgNQAAAAAAAAAAAAAAAAAA
-        AAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAA
-        AAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAB1c3RhciAgAGFyY2hpdmVtYXRpY2EAAAAAAAAA
-        AAAAAAAAAAAAAAAAYXJjaGl2ZW1hdGljYQAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAA
-        AAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAA
-        AAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAA
-        AAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAABt
-        YWtlX2RpcHNfMi03MjFiOThiOS1iODk0LTRjZmItODBhYi02MjRlNTIyNjMzMDAvcHJvY2Vzc2lu
-        Z01DUC54bWwAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAMDAwMDc3NgAwMDAwNTE1
-        ADAwMDA1MTUAMDAwMDAwMDUzMDMAMTMwMTAxNjUxNzMAMDI1MjczACAwAAAAAAAAAAAAAAAAAAAA
-        AAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAA
-        AAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAHVzdGFyICAAYXJjaGl2ZW1hdGljYQAAAAAAAAAA
-        AAAAAAAAAAAAAABhcmNoaXZlbWF0aWNhAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAA
-        AAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAA
-        AAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAA
-        AAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAADwh
-        LS0gQ29tbWVudHMgYXJlIG5vdCBhZGRlZCB3aGVuIHRoaXMgaXMgd3JpdHRlbiBieSB0aGUgYWRt
-        AAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAA=
-    headers:
-      Cache-Control: [no-cache]
-      Connection: [keep-alive]
-      Content-Length: ['235520']
-      Content-Type: [application/x-tar]
-      Date: ['Mon, 07 Nov 2016 21:56:13 GMT']
-      Server: [nginx/1.4.6 (Ubuntu)]
-      Vary: ['Accept, Cookie']
-      X-Frame-Options: [SAMEORIGIN]
-    status: {code: 200, message: OK}
-version: 1
