--- fixtures/vcr_cassettes/download_dip_no_dip.yaml
@@ -1,24 +1,0 @@
-interactions:
-- request:
-    body: null
-    headers:
-      Accept: ['*/*']
-      Accept-Encoding: ['gzip, deflate']
-      Connection: [keep-alive]
-      User-Agent: [python-requests/2.11.1]
-    method: GET
-    uri: http://192.168.168.192:8000/api/v2/file/bad%20dip%20uuid/download/?api_key=5de62f6f4817f903dcfac47fa5cffd44685a2cf2&username=test
-  response:
-    body:
-      string: !!binary |
-        H4sIAAAAAAAAAwMAAAAAAAAAAAA=
-    headers:
-      Connection: [keep-alive]
-      Content-Encoding: [gzip]
-      Content-Type: [text/html; charset=utf-8]
-      Date: ['Mon, 07 Nov 2016 22:06:56 GMT']
-      Server: [nginx/1.4.6 (Ubuntu)]
-      Vary: ['Accept, Cookie']
-      X-Frame-Options: [SAMEORIGIN]
-    status: {code: 404, message: NOT FOUND}
-version: 1
