--- fixtures/vcr_cassettes/get_package_details.yaml
@@ -1,53 +1,0 @@
-interactions:
-- request:
-    body: null
-    headers:
-      Accept: ['*/*']
-      Accept-Encoding: ['gzip, deflate']
-      Authorization: [!!python/unicode 'ApiKey test:test']
-      Connection: [keep-alive]
-      User-Agent: [python-requests/2.18.4]
-    method: GET
-    uri: http://192.168.168.192:8000/api/v2/file/23129471-09e3-467e-88b6-eb4714afb5ac
-  response:
-    body: {string: !!python/unicode ''}
-    headers:
-      connection: [keep-alive]
-      content-language: [en-us]
-      content-type: [text/html; charset=utf-8]
-      date: ['Sun, 15 Apr 2018 14:05:58 GMT']
-      location: ['http://127.0.0.1:62081/api/v2/file/23129471-09e3-467e-88b6-eb4714afb5ac/']
-      server: [nginx/1.12.2]
-      vary: [Accept-Language]
-      x-frame-options: [SAMEORIGIN]
-    status: {code: 301, message: MOVED PERMANENTLY}
-- request:
-    body: null
-    headers:
-      Accept: ['*/*']
-      Accept-Encoding: ['gzip, deflate']
-      Authorization: [!!python/unicode 'ApiKey test:test']
-      Connection: [keep-alive]
-      User-Agent: [python-requests/2.18.4]
-    method: GET
-    uri: http://127.0.0.1:62081/api/v2/file/23129471-09e3-467e-88b6-eb4714afb5ac/
-  response:
-    body: {string: !!python/unicode '{"current_full_path": "/var/archivematica/sharedDirectory/www/AIPsStore/2312/9471/09e3/467e/88b6/eb47/14af/b5ac/five-23129471-09e3-467e-88b6-eb4714afb5ac.7z",
-        "current_location": "/api/v2/location/b69754bb-3367-44f0-a00c-8eca0c0b53dd/",
-        "current_path": "2312/9471/09e3/467e/88b6/eb47/14af/b5ac/five-23129471-09e3-467e-88b6-eb4714afb5ac.7z",
-        "encrypted": false, "misc_attributes": {"reingest_pipeline": null}, "origin_pipeline":
-        "/api/v2/pipeline/cc3bf7dd-ab62-4e19-927e-6a5e196294e2/", "package_type":
-        "AIP", "related_packages": [], "replicas": [], "replicated_package": null,
-        "resource_uri": "/api/v2/file/23129471-09e3-467e-88b6-eb4714afb5ac/", "size":
-        19606, "status": "UPLOADED", "uuid": "23129471-09e3-467e-88b6-eb4714afb5ac"}'}
-    headers:
-      cache-control: [no-cache]
-      connection: [keep-alive]
-      content-language: [en]
-      content-type: [application/json]
-      date: ['Sun, 15 Apr 2018 14:05:58 GMT']
-      server: [nginx/1.12.2]
-      vary: ['Accept, Accept-Language, Cookie']
-      x-frame-options: [SAMEORIGIN]
-    status: {code: 200, message: OK}
-version: 1
