--- fixtures/vcr_cassettes/get_package_details_invalid_uuid.yaml
@@ -1,45 +1,0 @@
-interactions:
-- request:
-    body: null
-    headers:
-      Accept: ['*/*']
-      Accept-Encoding: ['gzip, deflate']
-      Authorization: [!!python/unicode 'ApiKey test:test']
-      Connection: [keep-alive]
-      User-Agent: [python-requests/2.18.4]
-    method: GET
-    uri: http://192.168.168.192:8000/api/v2/file/23129471-baad-f00d-88b6-eb4714afb5ac
-  response:
-    body: {string: !!python/unicode ''}
-    headers:
-      connection: [keep-alive]
-      content-language: [en-us]
-      content-type: [text/html; charset=utf-8]
-      date: ['Sun, 15 Apr 2018 14:03:13 GMT']
-      location: ['http://127.0.0.1:62081/api/v2/file/23129471-baad-f00d-88b6-eb4714afb5ac/']
-      server: [nginx/1.12.2]
-      vary: [Accept-Language]
-      x-frame-options: [SAMEORIGIN]
-    status: {code: 301, message: MOVED PERMANENTLY}
-- request:
-    body: null
-    headers:
-      Accept: ['*/*']
-      Accept-Encoding: ['gzip, deflate']
-      Authorization: [!!python/unicode 'ApiKey test:test']
-      Connection: [keep-alive]
-      User-Agent: [python-requests/2.18.4]
-    method: GET
-    uri: http://127.0.0.1:62081/api/v2/file/23129471-baad-f00d-88b6-eb4714afb5ac/
-  response:
-    body: {string: !!python/unicode ''}
-    headers:
-      connection: [keep-alive]
-      content-language: [en]
-      content-type: [text/html; charset=utf-8]
-      date: ['Sun, 15 Apr 2018 14:03:13 GMT']
-      server: [nginx/1.12.2]
-      vary: ['Accept, Accept-Language, Cookie']
-      x-frame-options: [SAMEORIGIN]
-    status: {code: 404, message: NOT FOUND}
-version: 1
