--- fixtures/vcr_cassettes/ingest_status_invalid_uuid.yaml
@@ -1,23 +1,0 @@
-interactions:
-- request:
-    body: null
-    headers:
-      Accept: ['*/*']
-      Accept-Encoding: ['gzip, deflate']
-      Authorization: [!!python/unicode 'ApiKey test:test']
-      Connection: [keep-alive]
-      User-Agent: [python-requests/2.18.4]
-    method: GET
-    uri: http://192.168.168.192/api/ingest/status/63fcc1b0-f83d-47e6-ac9d-a8f8d1fc2ab9/
-  response:
-    body: {string: !!python/unicode '{"message": "Cannot fetch unitSIP with UUID 63fcc1b0-f83d-47e6-ac9d-a8f8d1fc2ab9",
-        "type": "SIP", "error": true}'}
-    headers:
-      connection: [keep-alive]
-      content-language: [en]
-      content-type: [application/json]
-      date: ['Sun, 15 Apr 2018 13:41:55 GMT']
-      server: [nginx/1.12.2]
-      vary: ['Accept-Language, Cookie']
-    status: {code: 400, message: BAD REQUEST}
-version: 1
