--- fixtures/vcr_cassettes/pipeline_none.yaml
@@ -1,23 +1,0 @@
-interactions:
-- request:
-    body: null
-    headers:
-      Accept: ['*/*']
-      Accept-Encoding: ['gzip, deflate']
-      Connection: [keep-alive]
-      User-Agent: [python-requests/2.7.0 CPython/2.7.6 Linux/3.13.0-87-generic]
-      Authorization: ['ApiKey test:test']
-    method: GET
-    uri: http://192.168.168.192:8000/api/v2/pipeline/
-  response:
-    body: {string: !!python/unicode '{"meta": {"limit": 20, "next": null, 
-        "offset": 0, "previous": null, "total_count": 0}, "objects": []}'}
-    headers:
-      connection: [Keep-Alive]
-      content-type: [application/json]
-      date: ['Sat, 07 Apr 2018 16:59:25 GMT']
-      keep-alive: ['timeout=5, max=100']
-      server: [Apache/2.4.7 (Ubuntu)]
-      vary: [Cookie]
-    status: {code: 200, message: OK}
-version: 1
