--- fixtures/vcr_cassettes/reingest_exsting_aip.yaml
@@ -1,28 +1,0 @@
-interactions:
-- request:
-    body: !!python/unicode '{"pipeline": "65aaac5d-b4fd-478e-967b-6cdfee02f2c5", "reingest_type":
-      "full", "processing_config": "default"}'
-    headers:
-      Accept: ['*/*']
-      Accept-Encoding: ['gzip, deflate']
-      Authorization: [!!python/unicode 'ApiKey test:test']
-      Connection: [keep-alive]
-      Content-Length: ['109']
-      User-Agent: [python-requests/2.18.4]
-    method: POST
-    uri: http://192.168.168.192:8000/api/v2/file/df8e0c68-3bda-4d1d-8493-789f7dec47f5/reingest/
-  response:
-    body: {string: !!python/unicode '{"error": false, "message": "Package df8e0c68-3bda-4d1d-8493-789f7dec47f5
-        sent to pipeline Archivematica on 4e2f66a7a29f (65aaac5d-b4fd-478e-967b-6cdfee02f2c5)
-        for re-ingest", "reingest_uuid": "9dac7039-b0d8-4185-b27e-af008a9687ac", "status_code":
-        202}'}
-    headers:
-      connection: [keep-alive]
-      content-language: [en]
-      content-type: [application/json]
-      date: ['Sun, 08 Apr 2018 18:58:35 GMT']
-      server: [nginx/1.12.2]
-      vary: ['Accept, Accept-Language, Cookie']
-      x-frame-options: [SAMEORIGIN]
-    status: {code: 202, message: ACCEPTED}
-version: 1
