--- fixtures/vcr_cassettes/test_get_existing_processing_config.yaml
@@ -1,37 +1,0 @@
-interactions:
-- request:
-    body: null
-    headers:
-      Accept: ['*/*']
-      Accept-Encoding: ['gzip, deflate']
-      Authorization: [!!python/unicode 'ApiKey test:test']
-      Connection: [keep-alive]
-      User-Agent: [python-requests/2.18.4]
-    method: GET
-    uri: http://192.168.168.192/api/processing-configuration/default
-  response:
-    body: {string: !!python/unicode "<processingMCP>\n  <preconfiguredChoices>\n    <!--
-        Select compression level -->\n    <preconfiguredChoice>\n      <appliesTo>01c651cb-c174-4ba4-b985-1d87a44d6754</appliesTo>\n
-        \     <goToChain>ecfad581-b007-4612-a0e0-fcc551f4057f</goToChain>\n    </preconfiguredChoice>\n
-        \   <!-- Select file format identification command (Submission documentation
-        & metadata) -->\n    <preconfiguredChoice>\n      <appliesTo>087d27be-c719-47d8-9bbb-9a7d8b609c44</appliesTo>\n
-        \     <goToChain>25a91595-37f0-4373-a89a-56a757353fb8</goToChain>\n    </preconfiguredChoice>\n
-        \   <!-- Select compression algorithm -->\n    <preconfiguredChoice>\n      <appliesTo>01d64f58-8295-4b7b-9cab-8f1b153a504f</appliesTo>\n
-        \     <goToChain>9475447c-9889-430c-9477-6287a9574c5b</goToChain>\n    </preconfiguredChoice>\n
-        \   <!-- Delete packages after extraction -->\n    <preconfiguredChoice>\n
-        \     <appliesTo>f19926dd-8fb5-4c79-8ade-c83f61f55b40</appliesTo>\n      <goToChain>85b1e45d-8f98-4cae-8336-72f40e12cbef</goToChain>\n
-        \   </preconfiguredChoice>\n    <!-- Select file format identification command
-        (Ingest) -->\n    <preconfiguredChoice>\n      <appliesTo>7a024896-c4f7-4808-a240-44c87c762bc5</appliesTo>\n
-        \     <goToChain>3c1faec7-7e1e-4cdd-b3bd-e2f05f4baa9b</goToChain>\n    </preconfiguredChoice>\n
-        \   <!-- Send transfer to quarantine -->\n    <preconfiguredChoice>\n      <appliesTo>755b4177-c587-41a7-8c52-015277568302</appliesTo>\n
-        \     <goToChain>d4404ab1-dc7f-4e9e-b1f8-aa861e766b8e</goToChain>\n    </preconfiguredChoice>\n
-        \ </preconfiguredChoices>\n</processingMCP>\n"}
-    headers:
-      connection: [keep-alive]
-      content-language: [en]
-      content-type: [text/xml]
-      date: ['Sun, 08 Apr 2018 07:18:52 GMT']
-      server: [nginx/1.12.2]
-      vary: ['Accept-Language, Cookie']
-    status: {code: 200, message: OK}
-version: 1
