--- fixtures/vcr_cassettes/test_get_non_existing_processing_config.yaml
@@ -1,44 +1,0 @@
-interactions:
-- request:
-    body: null
-    headers:
-      Accept: ['*/*']
-      Accept-Encoding: ['gzip, deflate']
-      Authorization: [!!python/unicode 'ApiKey test:test']
-      Connection: [keep-alive]
-      User-Agent: [python-requests/2.18.4]
-    method: GET
-    uri: http://192.168.168.192/api/processing-configuration/badf00d
-  response:
-    body: {string: !!python/unicode "\n<!DOCTYPE html>\n<html lang=\"en\">\n<head>\n
-        \ <meta http-equiv=\"content-type\" content=\"text/html; charset=utf-8\">\n
-        \ <title>Page not found at /api/processing-configuration/badf00d</title>\n
-        \ <meta name=\"robots\" content=\"NONE,NOARCHIVE\">\n  <style type=\"text/css\">\n
-        \   html * { padding:0; margin:0; }\n    body * { padding:10px 20px; }\n    body
-        * * { padding:0; }\n    body { font:small sans-serif; background:#eee; }\n
-        \   body>div { border-bottom:1px solid #ddd; }\n    h1 { font-weight:normal;
-        margin-bottom:.4em; }\n    h1 span { font-size:60%; color:#666; font-weight:normal;
-        }\n    table { border:none; border-collapse: collapse; width:100%; }\n    td,
-        th { vertical-align:top; padding:2px 3px; }\n    th { width:12em; text-align:right;
-        color:#666; padding-right:.5em; }\n    #info { background:#f6f6f6; }\n    #info
-        ol { margin: 0.5em 4em; }\n    #info ol li { font-family: monospace; }\n    #summary
-        { background: #ffc; }\n    #explanation { background:#eee; border-bottom:
-        0px none; }\n  </style>\n</head>\n<body>\n  <div id=\"summary\">\n    <h1>Page
-        not found <span>(404)</span></h1>\n    <table class=\"meta\">\n      <tr>\n
-        \       <th>Request Method:</th>\n        <td>GET</td>\n      </tr>\n      <tr>\n
-        \       <th>Request URL:</th>\n        <td>http://127.0.0.1:62080/api/processing-configuration/badf00d</td>\n
-        \     </tr>\n      \n      <tr>\n        <th>Raised by:</th>\n        <td>components.api.views.wrapper</td>\n
-        \     </tr>\n      \n    </table>\n  </div>\n  <div id=\"info\">\n    \n      <p></p>\n
-        \   \n  </div>\n\n  <div id=\"explanation\">\n    <p>\n      You're seeing
-        this error because you have <code>DEBUG = True</code> in\n      your Django
-        settings file. Change that to <code>False</code>, and Django\n      will display
-        a standard 404 page.\n    </p>\n  </div>\n</body>\n</html>\n"}
-    headers:
-      connection: [keep-alive]
-      content-language: [en]
-      content-type: [text/html]
-      date: ['Sun, 08 Apr 2018 07:17:16 GMT']
-      server: [nginx/1.12.2]
-      vary: ['Accept-Language, Cookie']
-    status: {code: 404, message: NOT FOUND}
-version: 1
