--- fixtures/vcr_cassettes/test_hide_units.yaml
@@ -1,124 +1,0 @@
-interactions:
-- request:
-    body: null
-    headers:
-      Accept: ['*/*']
-      Accept-Encoding: ['gzip, deflate']
-      Connection: [keep-alive]
-      Content-Length: ['0']
-      User-Agent: [python-requests/2.18.1]
-    method: DELETE
-    uri: http://192.168.168.192/api/transfer/fdf1f7d4-7b0e-46d7-a1cc-e1851f8b92ed/delete/?username=test&api_key=3c23b0361887ace72b9d42963d9acbdf06644673
-  response:
-    body: {string: '{"removed": true}'}
-    headers:
-      Connection: [keep-alive]
-      Content-Language: [en]
-      Content-Type: [application/json]
-      Date: ['Mon, 07 May 2018 07:27:16 GMT']
-      Server: [nginx/1.12.2]
-      Vary: ['Accept-Language, Cookie']
-    status: {code: 200, message: OK}
-- request:
-    body: null
-    headers:
-      Accept: ['*/*']
-      Accept-Encoding: ['gzip, deflate']
-      Connection: [keep-alive]
-      Content-Length: ['0']
-      User-Agent: [python-requests/2.18.1]
-    method: DELETE
-    uri: http://192.168.168.192/api/transfer/777a9d9e-baad-f00d-8c7e-00b75773672d/delete/?username=test&api_key=3c23b0361887ace72b9d42963d9acbdf06644673
-  response:
-    body: {string: "\n<!DOCTYPE html>\n<html lang=\"en\">\n<head>\n  <meta http-equiv=\"content-type\"
-        content=\"text/html; charset=utf-8\">\n  <title>Page not found at /api/transfer/777a9d9e-baad-f00d-8c7e-00b75773672d/delete/</title>\n
-        \ <meta name=\"robots\" content=\"NONE,NOARCHIVE\">\n  <style type=\"text/css\">\n
-        \   html * { padding:0; margin:0; }\n    body * { padding:10px 20px; }\n    body
-        * * { padding:0; }\n    body { font:small sans-serif; background:#eee; }\n
-        \   body>div { border-bottom:1px solid #ddd; }\n    h1 { font-weight:normal;
-        margin-bottom:.4em; }\n    h1 span { font-size:60%; color:#666; font-weight:normal;
-        }\n    table { border:none; border-collapse: collapse; width:100%; }\n    td,
-        th { vertical-align:top; padding:2px 3px; }\n    th { width:12em; text-align:right;
-        color:#666; padding-right:.5em; }\n    #info { background:#f6f6f6; }\n    #info
-        ol { margin: 0.5em 4em; }\n    #info ol li { font-family: monospace; }\n    #summary
-        { background: #ffc; }\n    #explanation { background:#eee; border-bottom:
-        0px none; }\n  </style>\n</head>\n<body>\n  <div id=\"summary\">\n    <h1>Page
-        not found <span>(404)</span></h1>\n    <table class=\"meta\">\n      <tr>\n
-        \       <th>Request Method:</th>\n        <td>DELETE</td>\n      </tr>\n      <tr>\n
-        \       <th>Request URL:</th>\n        <td>http://192.168.168.192/api/transfer/777a9d9e-baad-f00d-8c7e-00b75773672d/delete/?username=test&amp;api_key=3c23b0361887ace72b9d42963d9acbdf06644673</td>\n
-        \     </tr>\n      \n      <tr>\n        <th>Raised by:</th>\n        <td>components.api.views.wrapper</td>\n
-        \     </tr>\n      \n    </table>\n  </div>\n  <div id=\"info\">\n    \n      <p></p>\n
-        \   \n  </div>\n\n  <div id=\"explanation\">\n    <p>\n      You're seeing
-        this error because you have <code>DEBUG = True</code> in\n      your Django
-        settings file. Change that to <code>False</code>, and Django\n      will display
-        a standard 404 page.\n    </p>\n  </div>\n</body>\n</html>\n"}
-    headers:
-      Connection: [keep-alive]
-      Content-Language: [en]
-      Content-Type: [text/html]
-      Date: ['Mon, 07 May 2018 07:27:16 GMT']
-      Server: [nginx/1.12.2]
-      Vary: ['Accept-Language, Cookie']
-    status: {code: 404, message: NOT FOUND}
-- request:
-    body: null
-    headers:
-      Accept: ['*/*']
-      Accept-Encoding: ['gzip, deflate']
-      Connection: [keep-alive]
-      Content-Length: ['0']
-      User-Agent: [python-requests/2.18.1]
-    method: DELETE
-    uri: http://192.168.168.192/api/ingest/b72afa68-9e82-410d-9235-02fa10512e14/delete/?username=test&api_key=3c23b0361887ace72b9d42963d9acbdf06644673
-  response:
-    body: {string: '{"removed": true}'}
-    headers:
-      Connection: [keep-alive]
-      Content-Language: [en]
-      Content-Type: [application/json]
-      Date: ['Mon, 07 May 2018 07:27:16 GMT']
-      Server: [nginx/1.12.2]
-      Vary: ['Accept-Language, Cookie']
-    status: {code: 200, message: OK}
-- request:
-    body: null
-    headers:
-      Accept: ['*/*']
-      Accept-Encoding: ['gzip, deflate']
-      Connection: [keep-alive]
-      Content-Length: ['0']
-      User-Agent: [python-requests/2.18.1]
-    method: DELETE
-    uri: http://192.168.168.192/api/ingest/777a9d9e-baad-f00d-8c7e-00b75773672d/delete/?username=test&api_key=3c23b0361887ace72b9d42963d9acbdf06644673
-  response:
-    body: {string: "\n<!DOCTYPE html>\n<html lang=\"en\">\n<head>\n  <meta http-equiv=\"content-type\"
-        content=\"text/html; charset=utf-8\">\n  <title>Page not found at /api/ingest/777a9d9e-baad-f00d-8c7e-00b75773672d/delete/</title>\n
-        \ <meta name=\"robots\" content=\"NONE,NOARCHIVE\">\n  <style type=\"text/css\">\n
-        \   html * { padding:0; margin:0; }\n    body * { padding:10px 20px; }\n    body
-        * * { padding:0; }\n    body { font:small sans-serif; background:#eee; }\n
-        \   body>div { border-bottom:1px solid #ddd; }\n    h1 { font-weight:normal;
-        margin-bottom:.4em; }\n    h1 span { font-size:60%; color:#666; font-weight:normal;
-        }\n    table { border:none; border-collapse: collapse; width:100%; }\n    td,
-        th { vertical-align:top; padding:2px 3px; }\n    th { width:12em; text-align:right;
-        color:#666; padding-right:.5em; }\n    #info { background:#f6f6f6; }\n    #info
-        ol { margin: 0.5em 4em; }\n    #info ol li { font-family: monospace; }\n    #summary
-        { background: #ffc; }\n    #explanation { background:#eee; border-bottom:
-        0px none; }\n  </style>\n</head>\n<body>\n  <div id=\"summary\">\n    <h1>Page
-        not found <span>(404)</span></h1>\n    <table class=\"meta\">\n      <tr>\n
-        \       <th>Request Method:</th>\n        <td>DELETE</td>\n      </tr>\n      <tr>\n
-        \       <th>Request URL:</th>\n        <td>http://192.168.168.192/api/ingest/777a9d9e-baad-f00d-8c7e-00b75773672d/delete/?username=test&amp;api_key=3c23b0361887ace72b9d42963d9acbdf06644673</td>\n
-        \     </tr>\n      \n      <tr>\n        <th>Raised by:</th>\n        <td>components.api.views.wrapper</td>\n
-        \     </tr>\n      \n    </table>\n  </div>\n  <div id=\"info\">\n    \n      <p></p>\n
-        \   \n  </div>\n\n  <div id=\"explanation\">\n    <p>\n      You're seeing
-        this error because you have <code>DEBUG = True</code> in\n      your Django
-        settings file. Change that to <code>False</code>, and Django\n      will display
-        a standard 404 page.\n    </p>\n  </div>\n</body>\n</html>\n"}
-    headers:
-      Connection: [keep-alive]
-      Content-Language: [en]
-      Content-Type: [text/html]
-      Date: ['Mon, 07 May 2018 07:27:16 GMT']
-      Server: [nginx/1.12.2]
-      Vary: ['Accept-Language, Cookie']
-    status: {code: 404, message: NOT FOUND}
-version: 1
