--- fixtures/vcr_cassettes/transfer_status.yaml
@@ -1,28 +1,0 @@
-interactions:
-- request:
-    body: null
-    headers:
-      Accept: ['*/*']
-      Accept-Encoding: ['gzip, deflate']
-      Connection: [keep-alive]
-      User-Agent: [python-requests/2.7.0 CPython/2.7.6 Linux/3.13.0-87-generic]
-      Authorization: ['ApiKey test:test']
-    method: GET
-    uri: http://192.168.168.192/api/transfer/status/63fcc1b0-f83d-47e6-ac9d-a8f8d1fc2ab9/
-  response:
-    body: {string: !!python/unicode '{"status": "COMPLETE", 
-        "name": "transfer_test_1", "sip_uuid": "BACKLOG", "microservice": 
-        "Create placement in backlog PREMIS events", "directory": 
-        "transfer_test_1-63fcc1b0-f83d-47e6-ac9d-a8f8d1fc2ab9", 
-        "path": "/var/archivematica/sharedDirectory/watchedDirectories/SIPCreation/completedTransfers/transfer_test_1-63fcc1b0-f83d-47e6-ac9d-a8f8d1fc2ab9/", 
-        "message": "Fetched status for 63fcc1b0-f83d-47e6-ac9d-a8f8d1fc2ab9 successfully.", 
-        "type": "transfer", "uuid": "63fcc1b0-f83d-47e6-ac9d-a8f8d1fc2ab9"}'}
-    headers:
-      connection: [Keep-Alive]
-      content-type: [application/json]
-      date: ['Sat, 07 Apr 2018 16:59:25 GMT']
-      keep-alive: ['timeout=5, max=100']
-      server: [Apache/2.4.7 (Ubuntu)]
-      vary: [Cookie]
-    status: {code: 200, message: OK}
-version: 1
