--- fixtures/vcr_cassettes/transferables_path.yaml
@@ -1,27 +1,0 @@
-interactions:
-- request:
-    body: null
-    headers:
-      Accept: ['*/*']
-      Accept-Encoding: ['gzip, deflate']
-      Connection: [keep-alive]
-      User-Agent: [python-requests/2.11.1]
-    method: GET
-    uri: http://192.168.168.192:8000/api/v2/location/7609101e-15b2-4f4f-a19d-7b23673ac93b/browse/?username=test&api_key=5de62f6f4817f903dcfac47fa5cffd44685a2cf2&path=dmFncmFudC9hcmNoaXZlbWF0aWNhLXNhbXBsZWRhdGE%3D
-  response:
-    body: {string: '{"directories": ["T1BGIGZvcm1hdC1jb3JwdXM=", "U2FtcGxlVHJhbnNmZXJz",
-        "VGVzdFRyYW5zZmVycw=="], "entries": ["T1BGIGZvcm1hdC1jb3JwdXM=", "UkVBRE1FLm1k",
-        "U2FtcGxlVHJhbnNmZXJz", "VGVzdFRyYW5zZmVycw=="], "properties": {"T1BGIGZvcm1hdC1jb3JwdXM=":
-        {"object count": 1499, "size": 4096}, "U2FtcGxlVHJhbnNmZXJz": {"object count":
-        95, "size": 4096}, "UkVBRE1FLm1k": {"size": 201}, "VGVzdFRyYW5zZmVycw==":
-        {"object count": 369, "size": 4096}}}'}
-    headers:
-      Cache-Control: [no-cache]
-      Connection: [keep-alive]
-      Content-Type: [application/json]
-      Date: ['Mon, 07 Nov 2016 19:44:26 GMT']
-      Server: [nginx/1.4.6 (Ubuntu)]
-      Vary: ['Accept, Cookie']
-      X-Frame-Options: [SAMEORIGIN]
-    status: {code: 200, message: OK}
-version: 1
