+++ requirements/base.txt
@@ -1,3 +1,4 @@
+amclient==1.0.0rc2
 metsrw==0.2.0
 requests<3.0
 sqlalchemy
