+++ transfers/transfer.py
@@ -27,8 +27,8 @@ from sqlalchemy.orm.exc import NoResultFound
 # by ensuring that it can see itself.
 sys.path.append(os.path.dirname(os.path.dirname(os.path.abspath(__file__))))
 
+from amclient import AMClient
 from transfers import defaults, errors, loggingconfig, models, utils
-from transfers.amclient import AMClient
 from transfers.transferargs import get_parser
 from transfers.utils import fsencode, fsdecode
